+++ example/lib/main.dart
@@ -81,10 +81,25 @@ class App extends StatelessWidget {
           theme: ShadThemeData(
             brightness: Brightness.light,
             colorScheme: const ShadZincColorScheme.light(),
+            // Example with google fonts
+            // textTheme: ShadTextTheme.fromGoogleFont(
+            //   from: GoogleFonts.pottaOne,
+            //   colorScheme: const ShadZincColorScheme.light(),
+            // ),
+            // Example of custom font family
+            // textTheme: ShadTextTheme(
+            //   colorScheme: const ShadZincColorScheme.light(),
+            //   family: 'UbuntuMono',
+            // ),
           ),
           darkTheme: ShadThemeData(
             brightness: Brightness.dark,
             colorScheme: const ShadZincColorScheme.dark(),
+            // Example of custom font family
+            // textTheme: ShadTextTheme(
+            //   colorScheme: const ShadZincColorScheme.dark(),
+            //   family: 'UbuntuMono',
+            // ),
           ),
           home: const MainPage(),
         );