+++ lib/shadcn_ui.dart
@@ -78,6 +78,8 @@ export 'src/theme/components/slider.dart';
 export 'src/theme/components/switch.dart';
 export 'src/theme/components/toast.dart';
 export 'src/theme/components/tooltip.dart';
+export 'src/theme/text_theme/theme.dart';
+export 'src/theme/text_theme/text_styles_default.dart';
 
 // Utils
 export 'src/utils/position.dart';