+++ lib/src/theme/text_theme/theme.dart
@@ -1,28 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:shadcn_ui/src/theme/text_theme/base.dart';
+import 'package:shadcn_ui/src/theme/color_scheme/base.dart';
+import 'package:shadcn_ui/src/theme/text_theme/text_styles_default.dart';
+import 'package:shadcn_ui/src/theme/themes/component_default.dart';
 
 @immutable
-class ShadTextThemeData extends ShadBaseTextTheme {
-  const ShadTextThemeData({
-    this.merge = true,
-    required super.h1Large,
-    required super.h1,
-    required super.h2,
-    required super.h3,
-    required super.h4,
-    required super.p,
-    required super.blockquote,
-    required super.table,
-    required super.list,
-    required super.lead,
-    required super.large,
-    required super.small,
-    required super.muted,
-  });
-
-  final bool merge;
-
-  ShadTextThemeData copyWith({
+class ShadTextTheme {
+  factory ShadTextTheme({
+    required ShadColorScheme colorScheme,
     TextStyle? h1Large,
     TextStyle? h1,
     TextStyle? h2,
@@ -36,8 +20,172 @@ class ShadTextThemeData extends ShadBaseTextTheme {
     TextStyle? large,
     TextStyle? small,
     TextStyle? muted,
+    String? family,
+    String? package,
   }) {
-    return ShadTextThemeData(
+    var effectiveFamily = family ?? kDefaultFontFamily;
+    if (package != null && package != '') {
+      effectiveFamily = 'packages/$package/$effectiveFamily';
+    }
+    return ShadTextTheme.custom(
+      h1Large: h1Large ??
+          ShadTextDefaultTheme.h1Large(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      h1: h1 ??
+          ShadTextDefaultTheme.h1(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      h2: h2 ??
+          ShadTextDefaultTheme.h2(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      h3: h3 ??
+          ShadTextDefaultTheme.h3(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      h4: h4 ??
+          ShadTextDefaultTheme.h4(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      p: p ??
+          ShadTextDefaultTheme.p(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      blockquote: blockquote ??
+          ShadTextDefaultTheme.blockquote(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      table: table ??
+          ShadTextDefaultTheme.table(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      list: list ??
+          ShadTextDefaultTheme.list(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      lead: lead ??
+          ShadTextDefaultTheme.lead(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      large: large ??
+          ShadTextDefaultTheme.large(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      small: small ??
+          ShadTextDefaultTheme.small(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      muted: muted ??
+          ShadTextDefaultTheme.muted(
+            colorScheme: colorScheme,
+            family: effectiveFamily,
+          ),
+      family: effectiveFamily,
+    );
+  }
+
+  const ShadTextTheme.custom({
+    this.merge = true,
+    required this.h1Large,
+    required this.h1,
+    required this.h2,
+    required this.h3,
+    required this.h4,
+    required this.p,
+    required this.blockquote,
+    required this.table,
+    required this.list,
+    required this.lead,
+    required this.large,
+    required this.small,
+    required this.muted,
+    required this.family,
+  });
+
+  factory ShadTextTheme.fromGoogleFont({
+    required TextStyle Function({
+      TextStyle? textStyle,
+    }) from,
+    required ShadColorScheme colorScheme,
+    ShadTextTheme? textTheme,
+  }) {
+    final effectiveTextTheme = textTheme ??
+        ShadComponentDefaultTheme.textTheme(colorScheme: colorScheme);
+    final h1Large = from(textStyle: effectiveTextTheme.h1Large);
+    return ShadTextTheme(
+      colorScheme: colorScheme,
+      h1Large: h1Large,
+      h1: from(textStyle: effectiveTextTheme.h1),
+      h2: from(textStyle: effectiveTextTheme.h2),
+      h3: from(textStyle: effectiveTextTheme.h3),
+      h4: from(textStyle: effectiveTextTheme.h4),
+      p: from(textStyle: effectiveTextTheme.p),
+      blockquote: from(textStyle: effectiveTextTheme.blockquote),
+      table: from(textStyle: effectiveTextTheme.table),
+      list: from(textStyle: effectiveTextTheme.list),
+      lead: from(textStyle: effectiveTextTheme.lead),
+      large: from(textStyle: effectiveTextTheme.large),
+      small: from(textStyle: effectiveTextTheme.small),
+      muted: from(textStyle: effectiveTextTheme.muted),
+      family: h1Large.fontFamily ?? effectiveTextTheme.family,
+    );
+  }
+
+  final TextStyle h1Large;
+  final TextStyle h1;
+  final TextStyle h2;
+  final TextStyle h3;
+  final TextStyle h4;
+  final TextStyle p;
+  final TextStyle blockquote;
+  final TextStyle table;
+  final TextStyle list;
+  final TextStyle lead;
+  final TextStyle large;
+  final TextStyle small;
+  final TextStyle muted;
+
+  /// The font family of the theme.
+  final String family;
+
+  final bool merge;
+
+  ShadTextTheme copyWith({
+    TextStyle? h1Large,
+    TextStyle? h1,
+    TextStyle? h2,
+    TextStyle? h3,
+    TextStyle? h4,
+    TextStyle? p,
+    TextStyle? blockquote,
+    TextStyle? table,
+    TextStyle? list,
+    TextStyle? lead,
+    TextStyle? large,
+    TextStyle? small,
+    TextStyle? muted,
+    String? family,
+    String? package,
+  }) {
+    var effectiveFamily = family ?? kDefaultFontFamily;
+    if (package != null && package != '') {
+      effectiveFamily = 'packages/$package/$effectiveFamily';
+    }
+
+    return ShadTextTheme.custom(
       merge: merge,
       h1Large: h1Large ?? this.h1Large,
       h1: h1 ?? this.h1,
@@ -52,12 +200,14 @@ class ShadTextThemeData extends ShadBaseTextTheme {
       large: large ?? this.large,
       small: small ?? this.small,
       muted: muted ?? this.muted,
+      family: effectiveFamily,
     );
   }
 
-  ShadTextThemeData mergeWith(ShadTextThemeData? other) {
+  ShadTextTheme mergeWith(ShadTextTheme? other) {
     if (other == null) return this;
     if (!other.merge) return other;
+
     return copyWith(
       h1Large: h1Large.merge(other.h1Large),
       h1: h1.merge(other.h1),
@@ -72,25 +222,28 @@ class ShadTextThemeData extends ShadBaseTextTheme {
       large: large.merge(other.large),
       small: small.merge(other.small),
       muted: muted.merge(other.muted),
+      family: other.family,
     );
   }
 
-  ShadTextThemeData apply({
-    String? fontFamily,
+  ShadTextTheme apply({
+    String? family,
     List<String>? fontFamilyFallback,
     String? package,
     double fontSizeFactor = 1.0,
     double fontSizeDelta = 0.0,
-    Color? displayColor, // Change color source
-    Color? bodyColor, // Change color source
-    Color? decorationColor, // Change color source
+    Color? displayColor,
+    Color? bodyColor,
+    Color? decorationColor,
     TextDecoration? decoration,
     TextDecorationStyle? decorationStyle,
   }) {
-    return ShadTextThemeData(
+    final effectiveFamily = family ?? this.family;
+    return ShadTextTheme.custom(
       merge: merge,
+      family: effectiveFamily,
       h1Large: h1Large.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -101,7 +254,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       h1: h1.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -112,7 +265,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       h2: h2.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -123,7 +276,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       h3: h3.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -134,7 +287,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       h4: h4.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -145,7 +298,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       p: p.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -156,7 +309,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       blockquote: blockquote.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -167,7 +320,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       table: table.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -178,7 +331,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       list: list.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -189,7 +342,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       lead: lead.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -200,7 +353,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         decorationStyle: decorationStyle,
       ),
       large: large.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -211,7 +364,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         color: bodyColor,
       ),
       small: small.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -222,7 +375,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         color: bodyColor,
       ),
       muted: muted.apply(
-        fontFamily: fontFamily,
+        fontFamily: effectiveFamily,
         fontFamilyFallback: fontFamilyFallback,
         package: package,
         fontSizeFactor: fontSizeFactor,
@@ -235,15 +388,15 @@ class ShadTextThemeData extends ShadBaseTextTheme {
     );
   }
 
-  static ShadTextThemeData lerp(
-    ShadTextThemeData? a,
-    ShadTextThemeData? b,
+  static ShadTextTheme lerp(
+    ShadTextTheme? a,
+    ShadTextTheme? b,
     double t,
   ) {
     if (identical(a, b) && a != null) {
       return a;
     }
-    return ShadTextThemeData(
+    return ShadTextTheme.custom(
       h1Large: TextStyle.lerp(a!.h1Large, b!.h1Large, t)!,
       h1: TextStyle.lerp(a.h1, b.h1, t)!,
       h2: TextStyle.lerp(a.h2, b.h2, t)!,
@@ -257,6 +410,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
       large: TextStyle.lerp(a.large, b.large, t)!,
       small: TextStyle.lerp(a.small, b.small, t)!,
       muted: TextStyle.lerp(a.muted, b.muted, t)!,
+      family: b.family,
     );
   }
 
@@ -268,7 +422,7 @@ class ShadTextThemeData extends ShadBaseTextTheme {
     if (other.runtimeType != runtimeType) {
       return false;
     }
-    return other is ShadTextThemeData &&
+    return other is ShadTextTheme &&
         other.h1Large == h1Large &&
         other.h1 == h1 &&
         other.h2 == h2 &&
@@ -281,7 +435,8 @@ class ShadTextThemeData extends ShadBaseTextTheme {
         other.lead == lead &&
         other.large == large &&
         other.small == small &&
-        other.muted == muted;
+        other.muted == muted &&
+        other.family == family;
   }
 
   @override
@@ -300,6 +455,8 @@ class ShadTextThemeData extends ShadBaseTextTheme {
       large,
       small,
       muted,
+      family,
+      merge,
     );
   }
 }