+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.3
+
+- Expose `ShadTextTheme`
+- Fix `ShadTextTheme` to add ability to change font family
+- Add `materialThemeBuilder` to default `ShadApp`
+- Add `ShadTextTheme.fromGoogleFont` to use a Google Font
+
 ## 0.2.2
 
 - Fix lerp of `ShadDecoration` and `ShadBorder`